break out Edit Profile buttons into seperate functions
